+++ src/scripts/card.js
@@ -1,4 +1,8 @@
-import {api} from './api.js';
+//import {api} from './api.js';
+import {Popup} from './popup.js';
+const placesList = document.querySelector('.places-list');
+const popupImage = new Popup(document.querySelector('.popup'));
+
 export class Card {
   constructor(name, link){
   this.cardElement = this.create(name, link);
