+++ src/scripts/cardList.js
@@ -1,5 +1,5 @@
 import {Card} from './card.js';
-import {api} from './api.js';
+
 
 export class CardList{
   constructor(container, array){
